discovery: consolidate constructors into single file
